+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * drop support for R 4.2 due to changes in dependency packages;
 * use `PredictiveEcology/SpaDES.project@development` following merge of `transition` branch;
+* add `from_wd` argument to `findProjectPath`;
 * documentation improvements;
 * update and add new tests;